Restructuring the searchspace file structure
