+++ continuousresource/localsearch/__init__.py
@@ -1,3 +1,4 @@
-from .searchspace import *  # noqa
 from .localsearch import *  # noqa
+from .searchspace_abstract import *  # noqa
+from .searchspace_jobarray import *  # noqa
 from .utils import *  # noqa
